Выполнен рефакторинг тестов Entity Framework.
